refactor(rules): extract AddJavascriptInterfaceRule from android_security.go
android_security.go was 2570 lines covering ~20 distinct rules. Step
one of a god-file split: move AddJavascriptInterfaceRule and its 11
addJavascriptInterface*/ancestorDirs helpers into
android_security_javascript_interface.go (276 LOC). The group is
fully self-contained — every helper is called only from this rule.

Brings the main file to 2294 LOC. Follow-up PRs will extract the
crypto and component-export groups; those have shared helpers and
need separate split planning.
